Algo: optimize algorithm code structure and add model training scripts
